Updated Physician

+++ src/main/java/com/example/hms/service/PhysicianService.java
@@ -1,6 +1,6 @@
 package com.example.hms.service;
 
-
+import com.example.hms.entity.Nurse;
 import com.example.hms.entity.Physician;
 import com.example.hms.repository.PhysicianRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,18 +24,40 @@ public class PhysicianService {
         return physicianRepository.findById(id).orElse(null);
     }
 
+    //get physician by name
+    public List<Physician> getPhysicianByName(String name){
+
+        return physicianRepository.findByName(name);
+    }
+
+    //get physician by position
+    public List<Physician> getPhysicianByPosition(String position){
+
+        return physicianRepository.findByPosition(position);
+    }
+
     public Physician addPhysician(Physician physician) {
 
         return physicianRepository.save(physician);
     }
 
-    public Physician updatePhysician(int id, Physician physician) {
-        physician.setEmployeeId(id);
-        return physicianRepository.save(physician);
+    //Update Existing Physician
+    public Physician updatePhysician(int id, Physician updatedPhysician) {
+
+        Physician existingPhysician = physicianRepository.findById(id).orElseThrow(() -> new RuntimeException("Physician not found"));
+
+        existingPhysician.setName(updatedPhysician.getName());
+        existingPhysician.setPosition(updatedPhysician.getPosition());
+        existingPhysician.setSsn(updatedPhysician.getSsn());
+
+        return physicianRepository.save(existingPhysician);
     }
 
     public void deletePhysician(int id) {
 
+        if(!physicianRepository.existsById(id)){
+            throw new RuntimeException("Physician not found."+id);
+        }
         physicianRepository.deleteById(id);
     }
 }
